@@ -1,5 +1,4 @@
-import _Vue = require("vue");
-import { WatchOptions } from "vue";
+import _Vue, { WatchOptions } from "vue";
 
 // augment typings of Vue.js
 import "./vue";
@@ -17,11 +16,12 @@ export declare class Store<S> {
   dispatch: Dispatch;
   commit: Commit;
 
-  subscribe<P extends Payload>(fn: (mutation: P, state: S) => any): () => void;
-  watch<T>(getter: (state: S) => T, cb: (value: T, oldValue: T) => void, options?: WatchOptions): void;
+  subscribe<P extends MutationPayload>(fn: (mutation: P, state: S) => any): () => void;
+  subscribeAction<P extends ActionPayload>(fn: (action: P, state: S) => any): () => void;
+  watch<T>(getter: (state: S, getters: any) => T, cb: (value: T, oldValue: T) => void, options?: WatchOptions): () => void;
 
-  registerModule<T>(path: string, module: Module<T, S>): void;
-  registerModule<T>(path: string[], module: Module<T, S>): void;
+  registerModule<T>(path: string, module: Module<T, S>, options?: ModuleOptions): void;
+  registerModule<T>(path: string[], module: Module<T, S>, options?: ModuleOptions): void;
 
   unregisterModule(path: string): void;
   unregisterModule(path: string[]): void;
@@ -59,6 +59,14 @@ export interface Payload {
   type: string;
 }
 
+export interface MutationPayload extends Payload {
+  payload: any;
+}
+
+export interface ActionPayload extends Payload {
+  payload: any;
+}
+
 export interface DispatchOptions {
   root?: boolean;
 }
@@ -78,8 +86,14 @@ export interface StoreOptions<S> {
   strict?: boolean;
 }
 
+type ActionHandler<S, R> = (injectee: ActionContext<S, R>, payload: any) => any;
+interface ActionObject<S, R> {
+  root?: boolean;
+  handler: ActionHandler<S, R>;
+}
+
 export type Getter<S, R> = (state: S, getters: any, rootState: R, rootGetters: any) => any;
-export type Action<S, R> = (injectee: ActionContext<S, R>, payload: any) => any;
+export type Action<S, R> = ActionHandler<S, R> | ActionObject<S, R>;
 export type Mutation<S> = (state: S, payload: any) => any;
 export type Plugin<S> = (store: Store<S>) => any;
 
@@ -90,6 +104,10 @@ export interface Module<S, R> {
   actions?: ActionTree<S, R>;
   mutations?: MutationTree<S>;
   modules?: ModuleTree<R>;
+}
+
+export interface ModuleOptions{
+  preserveState?: boolean
 }
 
 export interface GetterTree<S, R> {
@@ -107,3 +125,9 @@ export interface MutationTree<S> {
 export interface ModuleTree<R> {
   [key: string]: Module<any, R>;
 }
+
+declare const _default: {
+  Store: typeof Store;
+  install: typeof install;
+}
+export default _default;
