@@ -27,7 +27,7 @@ const mutations = {
   },
 
   [types.ADD_TO_CART] (state, { id }) {
-    state.all.find(p => p.id === id).inventory--
+    state.all.find(product => product.id === id).inventory--
   }
 }
 
