@@ -2,6 +2,7 @@ import Vue from 'vue/dist/vue.common.js'
 import Vuex from '../../dist/vuex.common.js'
 
 const TEST = 'TEST'
+const isSSR = process.env.VUE_ENV === 'server'
 
 describe('Store', () => {
   it('committing mutations', () => {
@@ -263,77 +264,6 @@ describe('Store', () => {
     )
   })
 
-  it('strict mode: warn mutations outside of handlers', () => {
-    const store = new Vuex.Store({
-      state: {
-        a: 1
-      },
-      strict: true
-    })
-    Vue.config.silent = true
-    expect(() => { store.state.a++ }).toThrow()
-    Vue.config.silent = false
-  })
-
-  it('watch: with resetting vm', done => {
-    const store = new Vuex.Store({
-      state: {
-        count: 0
-      },
-      mutations: {
-        [TEST]: state => state.count++
-      }
-    })
-
-    const spy = jasmine.createSpy()
-    store.watch(state => state.count, spy)
-
-    // reset store vm
-    store.registerModule('test', {})
-
-    Vue.nextTick(() => {
-      store.commit(TEST)
-      expect(store.state.count).toBe(1)
-
-      Vue.nextTick(() => {
-        expect(spy).toHaveBeenCalled()
-        done()
-      })
-    })
-  })
-
-  it('watch: getter function has access to store\'s getters object', done => {
-    const store = new Vuex.Store({
-      state: {
-        count: 0
-      },
-      mutations: {
-        [TEST]: state => state.count++
-      },
-      getters: {
-        getCount: state => state.count
-      }
-    })
-
-    const getter = function getter (state, getters) {
-      return state.count
-    }
-    const spy = spyOn({ getter }, 'getter').and.callThrough()
-    const spyCb = jasmine.createSpy()
-
-    store.watch(spy, spyCb)
-
-    Vue.nextTick(() => {
-      store.commit(TEST)
-      expect(store.state.count).toBe(1)
-
-      Vue.nextTick(() => {
-        expect(spy).toHaveBeenCalledWith(store.state, store.getters)
-        done()
-      })
-    })
-  })
-
   it('asserts the call with the new operator', () => {
     expect(() => {
       Vuex.Store({})
@@ -355,4 +285,112 @@ describe('Store', () => {
     store.commit(TEST, 2)
     expect(store.state.a).toBe(3)
   })
+
+  it('should not call root state function twice', () => {
+    const spy = jasmine.createSpy().and.returnValue(1)
+    new Vuex.Store({
+      state: spy
+    })
+    expect(spy).toHaveBeenCalledTimes(1)
+  })
+
+  it('subscribe: should handle subscriptions / unsubscriptions', () => {
+    const subscribeSpy = jasmine.createSpy()
+    const secondSubscribeSpy = jasmine.createSpy()
+    const testPayload = 2
+    const store = new Vuex.Store({
+      state: {},
+      mutations: {
+        [TEST]: () => {}
+      }
+    })
+
+    const unsubscribe = store.subscribe(subscribeSpy)
+    store.subscribe(secondSubscribeSpy)
+    store.commit(TEST, testPayload)
+    unsubscribe()
+    store.commit(TEST, testPayload)
+
+    expect(subscribeSpy).toHaveBeenCalledWith(
+      { type: TEST, payload: testPayload },
+      store.state
+    )
+    expect(secondSubscribeSpy).toHaveBeenCalled()
+    expect(subscribeSpy.calls.count()).toBe(1)
+    expect(secondSubscribeSpy.calls.count()).toBe(2)
+  })
+
+  // store.watch should only be asserted in non-SSR environment
+  if (!isSSR) {
+    it('strict mode: warn mutations outside of handlers', () => {
+      const store = new Vuex.Store({
+        state: {
+          a: 1
+        },
+        strict: true
+      })
+      Vue.config.silent = true
+      expect(() => { store.state.a++ }).toThrow()
+      Vue.config.silent = false
+    })
+
+    it('watch: with resetting vm', done => {
+      const store = new Vuex.Store({
+        state: {
+          count: 0
+        },
+        mutations: {
+          [TEST]: state => state.count++
+        }
+      })
+
+      const spy = jasmine.createSpy()
+      store.watch(state => state.count, spy)
+
+      // reset store vm
+      store.registerModule('test', {})
+
+      Vue.nextTick(() => {
+        store.commit(TEST)
+        expect(store.state.count).toBe(1)
+
+        Vue.nextTick(() => {
+          expect(spy).toHaveBeenCalled()
+          done()
+        })
+      })
+    })
+
+    it('watch: getter function has access to store\'s getters object', done => {
+      const store = new Vuex.Store({
+        state: {
+          count: 0
+        },
+        mutations: {
+          [TEST]: state => state.count++
+        },
+        getters: {
+          getCount: state => state.count
+        }
+      })
+
+      const getter = function getter (state, getters) {
+        return state.count
+      }
+      const spy = spyOn({ getter }, 'getter').and.callThrough()
+      const spyCb = jasmine.createSpy()
+
+      store.watch(spy, spyCb)
+
+      Vue.nextTick(() => {
+        store.commit(TEST)
+        expect(store.state.count).toBe(1)
+
+        Vue.nextTick(() => {
+          expect(spy).toHaveBeenCalledWith(store.state, store.getters)
+          done()
+        })
+      })
+    })
+  }
 })
